dataBaseClient tests pass
